Move PrunedLiveness so it can be used as a lightweight OSSA helper.
For use in OwnershipUtils.
